+++ trin-history/Cargo.toml
@@ -18,3 +18,4 @@ tracing = "0.1.26"
 tracing-subscriber = "0.2.18"
 tokio = { version = "1.8.0", features = ["full"] }
 trin-core = { path = "../trin-core" }
+parking_lot = "0.11.1"
