+++ trin-history/src/lib.rs
@@ -4,9 +4,11 @@ use tokio::sync::mpsc;
 
 use trin_core::cli::TrinConfig;
 use trin_core::portalnet::discovery::Discovery;
-use trin_core::portalnet::protocol::{
-    HistoryEndpointKind, HistoryNetworkEndpoint, PortalnetConfig, PortalnetProtocol,
-};
+use trin_core::portalnet::protocol::{HistoryEndpointKind, HistoryNetworkEndpoint, PortalnetConfig, PortalnetEvents};
+use trin_core::portalnet::overlay::OverlayProtocol;
+use trin_core::utils::setup_overlay_db;
+use std::sync::Arc;
+use parking_lot::RwLock;
 
 pub struct HistoryRequestHandler {
     pub history_rx: mpsc::UnboundedReceiver<HistoryNetworkEndpoint>,
@@ -67,15 +69,29 @@ pub async fn main() -> Result<(), Box<dyn std::error::Error>> {
         portalnet_config.bootnode_enrs
     );
 
-    tokio::spawn(async move {
-        let (mut p2p, events) = PortalnetProtocol::new(discovery, portalnet_config)
+    let protocol_receiver = discovery
+            .discv5
+            .event_stream()
             .await
-            .unwrap();
+            .map_err(|e| e.to_string()).unwrap();
+
+    tokio::spawn(async move {
+        let discovery = Arc::new(discovery);
+        let overlay = OverlayProtocol::new(discovery.clone(),portalnet_config.clone());
+        let db = setup_overlay_db(discovery.local_enr());
+
+        let events = PortalnetEvents {
+            discovery,
+            overlay,
+            protocol_receiver,
+            db,
+        };
+        // let events = PortalnetEvents::new(discovery, overlay.clone(), db).await;
 
         tokio::spawn(events.process_discv5_requests());
 
         // hacky test: make sure we establish a session with the boot node
-        p2p.ping_bootnodes().await.unwrap();
+        // overlay.ping_bootnodes().await.unwrap();
 
         tokio::signal::ctrl_c()
             .await
