+++ trin-state/src/lib.rs
@@ -4,9 +4,10 @@ use tokio::sync::mpsc;
 
 use trin_core::cli::TrinConfig;
 use trin_core::portalnet::discovery::Discovery;
-use trin_core::portalnet::protocol::{
-    PortalnetConfig, PortalnetProtocol, StateEndpointKind, StateNetworkEndpoint,
-};
+use trin_core::portalnet::protocol::{PortalnetConfig, StateEndpointKind, StateNetworkEndpoint, PortalnetEvents};
+use trin_core::portalnet::overlay::OverlayProtocol;
+use trin_core::utils::setup_overlay_db;
+use std::sync::Arc;
 
 pub mod utils;
 
@@ -57,7 +58,7 @@ pub async fn main() -> Result<(), Box<dyn std::error::Error>> {
         ..Default::default()
     };
 
-    let mut discovery = Discovery::new(portalnet_config.clone()).unwrap();
+    let mut discovery = Arc::new(Discovery::new(portalnet_config.clone()).unwrap());
     discovery.start().await.unwrap();
 
     info!(
@@ -66,14 +67,14 @@ pub async fn main() -> Result<(), Box<dyn std::error::Error>> {
     );
 
     tokio::spawn(async move {
-        let (mut p2p, events) = PortalnetProtocol::new(discovery, portalnet_config)
-            .await
-            .unwrap();
+        let mut overlay = OverlayProtocol::new(discovery.clone(),portalnet_config.clone());
+        let db = setup_overlay_db(discovery.local_enr());
+        let events = PortalnetEvents::new(discovery, overlay.clone(), db).await;
 
         tokio::spawn(events.process_discv5_requests());
 
         // hacky test: make sure we establish a session with the boot node
-        p2p.ping_bootnodes().await.unwrap();
+        overlay.ping_bootnodes().await.unwrap();
 
         tokio::signal::ctrl_c()
             .await
